+++ plan.php
@@ -526,6 +526,15 @@ function dict(array $structure, $required=false, $extra=false)
     };
 }
 
+function file()
+{
+    return assert\dict(
+        ['error' => 0],
+        ['tmp_name', 'size', 'error', 'name', 'type'],
+        false
+    );
+}
+
 /**
  * Runs a validator through a list of data keys.
  *